Separate driver to cli file
