+++ markdown_lt/__init__.py
@@ -1,7 +1,8 @@
 """LanguageTool for markdown files."""
 
-__all__ = ['AstMatcher', 'AstRenderer', 'Linter']
+__all__ = ['run', 'AstMatcher', 'AstRenderer', 'Linter']
 
+from markdown_lt.cli import run
 from markdown_lt.astmatcher import AstMatcher
 from markdown_lt.ast import AstRenderer
 from markdown_lt.linter import Linter
